A bit of refactoring on test project
