+++ tests/CustomCode-Analyzer.Tests/AnalyzerTests.cs
@@ -1,179 +1,14 @@
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.Testing;
-using Microsoft.CodeAnalysis.CSharp.Testing;
-using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Microsoft.CodeAnalysis.Testing.Verifiers;
 using System.Threading.Tasks;
 using static Analyzer;
-using System.Collections.Immutable;
-using System.Text;
-using Xunit.Abstractions;
 
 namespace CustomCode_Analyzer.Tests
 {
-    /// <summary>
-    /// Provides detailed logging capabilities for analyzer tests.
-    /// Implements ITestOutputHelper to capture and display test execution details.
-    /// </summary>
-    public class DetailedTestLogger : ITestOutputHelper
-    {
-        private readonly TestContext _testContext;
-        private readonly StringBuilder _output = new();
-
-        public DetailedTestLogger(TestContext testContext)
-        {
-            _testContext = testContext;
-        }
-
-        public void WriteLine(string message)
-        {
-            _output.AppendLine(message);
-            _testContext.WriteLine(message);
-        }
-
-        public void WriteLine(string format, params object[] args)
-        {
-            WriteLine(string.Format(format, args));
-        }
-
-        public void Clear() => _output.Clear();
-
-        public override string ToString() => _output.ToString();
-    }
-
-    // Generic verifier class for analyzer tests
-    public static class CSharpAnalyzerVerifier<TAnalyzer>
-        where TAnalyzer : DiagnosticAnalyzer, new()
-    {
-        // Create a diagnostic result for the specified rule
-        public static DiagnosticResult Diagnostic(string diagnosticId)
-            => CSharpAnalyzerVerifier<TAnalyzer, DefaultVerifier>.Diagnostic(diagnosticId);
-
-        // Main method to verify analyzer behavior against source code
-        public static async Task VerifyAnalyzerAsync(string source, TestContext testContext, params DiagnosticResult[] expected)
-        {
-            var logger = new DetailedTestLogger(testContext);
-            var test = new Test(logger) { TestCode = source };
-
-            logger.WriteLine($"Running test: {testContext.TestName}");
-            logger.WriteLine("\nAnalyzing source code:");
-            logger.WriteLine(source);
-
-            // Add required attribute definitions that would normally be in referenced assemblies
-            test.TestState.Sources.Add(@"
-        using System;
-        
-        [AttributeUsage(AttributeTargets.Interface)]
-        public class OSInterfaceAttribute : Attribute 
-        { 
-            public string Name { get; set; }
-            public string Description { get; set; }
-            public string IconResourceName { get; set; }
-            public string OriginalName { get; set; }
-        }
-        
-        [AttributeUsage(AttributeTargets.Struct)]
-        public class OSStructureAttribute : Attribute 
-        { 
-            public string Name { get; set; }
-            public string Description { get; set; }
-            public string OriginalName { get; set; }
-        }
-        
-        [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
-        public class OSStructureFieldAttribute : Attribute 
-        {
-            public string Description { get; set; }
-            public int Length { get; set; } = 50;
-            public string OriginalName { get; set; }
-        }
-        
-        [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
-        public class OSIgnoreAttribute : Attribute { }
-    ");
-            // Add expected diagnostics to test
-            test.ExpectedDiagnostics.AddRange(expected);
-
-            // Log expected diagnostics for debugging
-            logger.WriteLine("\nExpected diagnostics:");
-            foreach (var diagnostic in expected)
-            {
-                logger.WriteLine($"- {diagnostic.Id}: {diagnostic.MessageFormat} at {diagnostic.Spans[0]}");
-            }
-
-            // Run the analyzer test
-            await test.RunAsync();
-        }
-
-        // Internal test class that configures and runs individual analyzer tests
-        private class Test : CSharpAnalyzerTest<TAnalyzer, DefaultVerifier>
-        {
-            private readonly ITestOutputHelper _logger;
-
-            public Test(ITestOutputHelper logger)
-            {
-                // Configure test to use .NET 8.0 reference assemblies
-                _logger = logger;
-                ReferenceAssemblies = new ReferenceAssemblies(
-                    "net8.0",
-                    new PackageIdentity(
-                        "Microsoft.NETCore.App.Ref",
-                        "8.0.0"),
-                    Path.Combine("ref", "net8.0"));
-
-                // Add solution-wide configuration for nullable warnings
-                SolutionTransforms.Add((solution, projectId) =>
-                {
-                    var compilationOptions = solution.GetProject(projectId)?.CompilationOptions;
-                    if (compilationOptions != null)
-                    {
-                        // Configure specific diagnostic options for nullable warnings
-                        compilationOptions = compilationOptions.WithSpecificDiagnosticOptions(
-                            compilationOptions.SpecificDiagnosticOptions.SetItems(CSharpVerifierHelper.NullableWarnings));
-                        solution = solution.WithProjectCompilationOptions(projectId, compilationOptions);
-                    }
-                    return solution;
-                });
-            }
-
-            // Override to configure specific compilation options for each test
-            protected override CompilationOptions CreateCompilationOptions()
-            {
-                var compilationOptions = base.CreateCompilationOptions();
-                return compilationOptions.WithSpecificDiagnosticOptions(
-                    compilationOptions.SpecificDiagnosticOptions.SetItems(CSharpVerifierHelper.NullableWarnings));
-            }
-        }
-    }
-
-    // Helper class to configure nullable reference type warnings
-    public static class CSharpVerifierHelper
-    {
-        // Dictionary of nullable warning configurations
-        public static ImmutableDictionary<string, ReportDiagnostic> NullableWarnings
-        {
-            get
-            {
-                return ImmutableDictionary.CreateRange(new[]
-                {
-                        // Configure specific nullable warning codes as errors
-                        new KeyValuePair<string, ReportDiagnostic>("CS8632", ReportDiagnostic.Error),
-                        new KeyValuePair<string, ReportDiagnostic>("CS8669", ReportDiagnostic.Error)
-                    });
-            }
-        }
-    }
-
-    // Define OSInterface attribute for test use
-    [AttributeUsage(AttributeTargets.Interface)]
-    public class OSInterfaceAttribute : Attribute { }
-
     [TestClass]
     public class AnalyzerTests
     {
         public TestContext? TestContext { get; set; }
-
 
         [TestMethod]
         public async Task NonPublicOSInterface_ReportsWarning()
