+++ tests/CustomCode-Analyzer.Tests/CSharpVerifierHelper.cs
@@ -1,0 +1,23 @@
+﻿using Microsoft.CodeAnalysis;
+using System.Collections.Immutable;
+
+namespace CustomCode_Analyzer.Tests
+{
+    // Helper class to configure nullable reference type warnings
+    public static class CSharpVerifierHelper
+    {
+        // Dictionary of nullable warning configurations
+        public static ImmutableDictionary<string, ReportDiagnostic> NullableWarnings
+        {
+            get
+            {
+                return ImmutableDictionary.CreateRange(new[]
+                {
+                        // Configure specific nullable warning codes as errors
+                        new KeyValuePair<string, ReportDiagnostic>("CS8632", ReportDiagnostic.Error),
+                        new KeyValuePair<string, ReportDiagnostic>("CS8669", ReportDiagnostic.Error)
+                    });
+            }
+        }
+    }
+}
