+++ tests/CustomCode-Analyzer.Tests/CustomCode-Analyzer.Tests.csproj
@@ -26,6 +26,7 @@
 		  <PrivateAssets>all</PrivateAssets>
 		  <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 		</PackageReference>
+		<PackageReference Include="OutSystems.ExternalLibraries.SDK" Version="1.5.0" />
 		<PackageReference Include="xunit.abstractions" Version="2.0.3" />
 	</ItemGroup>
